moved plots from use cases to pymodule plot function
